export all core classes in core:module
all dependencies
- provider
- flex_color_scheme
- core:module

update analysis_options.yaml
- add riles
