+++ core/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: db747aa1331bd95bc9b3874c842261ca2d302cd5
+  channel: stable
+
+project_type: package
