+++ core/lib/exceptions/invalid_message_exception.dart
@@ -1,0 +1,8 @@
+class InvalidMessageException implements Exception {
+  final String message;
+
+  const InvalidMessageException(this.message);
+
+  @override
+  String toString() => 'InvalidMessageException(message: $message)';
+}
