+++ core/lib/exceptions/invalid_name_exception.dart
@@ -1,0 +1,8 @@
+class InvalidNameException implements Exception {
+  final String message;
+
+  const InvalidNameException(this.message);
+
+  @override
+  String toString() => 'InvalidNameException(message: $message)';
+}
