+++ core/lib/models/dtos/geolocation.dart
