+++ core/lib/models/dtos/user/chat_user_dto.dart
@@ -1,0 +1,11 @@
+/// Basic model representing chat user.
+class ChatUserDto {
+  final String name;
+
+  const ChatUserDto({
+    required this.name,
+  });
+
+  @override
+  String toString() => 'ChatUserDto(name: $name)';
+}
