+++ core/lib/models/dtos/user/chat_user_local_dto.dart
@@ -1,14 +1,4 @@
-/// Basic model representing chat user.
-class ChatUserDto {
-  final String name;
-
-  const ChatUserDto({
-    required this.name,
-  });
-
-  @override
-  String toString() => 'ChatUserDto(name: $name)';
-}
+import 'chat_user_dto.dart';
 
 /// Model representing local user.
 ///
