--- lib/data/chat/chat.dart
@@ -1,4 +1,0 @@
-export 'repository/repository.dart';
-export 'models/message.dart';
-export 'models/user.dart';
-export 'models/geolocation.dart';
