+++ lib/data/chat/repository/firebase.dart
@@ -1,9 +1,13 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-
-import 'package:surf_practice_chat_flutter/data/chat/models/geolocation.dart';
-import 'package:surf_practice_chat_flutter/data/chat/models/message.dart';
-import 'package:surf_practice_chat_flutter/data/chat/models/user.dart';
-import 'package:surf_practice_chat_flutter/data/chat/repository/repository.dart';
+import 'package:core/contracts/repositories/chat_repository.dart';
+import 'package:core/exceptions/invalid_message_exception.dart';
+import 'package:core/exceptions/invalid_name_exception.dart';
+import 'package:core/models/dtos/firebase/message_firebase_dto.dart';
+import 'package:core/models/dtos/geolocation.dart';
+import 'package:core/models/dtos/message/chat_message_dto.dart';
+import 'package:core/models/dtos/message/chat_message_geolocation_dto.dart';
+import 'package:core/models/dtos/user/chat_user_dto.dart';
+import 'package:core/models/dtos/user/chat_user_local_dto.dart';
 
 class ChatRepositoryFirebase implements ChatRepository {
   static const String _messagesCollectionKey = 'messages';
@@ -37,9 +41,9 @@ class ChatRepositoryFirebase implements ChatRepository {
     _savedLocalName = nickname;
 
     await _firebaseClient.collection(_messagesCollectionKey).add({
-      _MessageFirebaseDto._authorNameKey: nickname,
-      _MessageFirebaseDto._messageKey: message,
-      _MessageFirebaseDto._createdKey: FieldValue.serverTimestamp(),
+      MessageFirebaseDto.authorNameKey: nickname,
+      MessageFirebaseDto.messageKey: message,
+      MessageFirebaseDto.createdKey: FieldValue.serverTimestamp(),
     });
 
     return messages;
@@ -57,10 +61,10 @@ class ChatRepositoryFirebase implements ChatRepository {
     _savedLocalName = nickname;
 
     await _firebaseClient.collection(_messagesCollectionKey).add({
-      _MessageFirebaseDto._authorNameKey: nickname,
-      _MessageFirebaseDto._messageKey: message ?? '',
-      _MessageFirebaseDto._createdKey: FieldValue.serverTimestamp(),
-      _MessageFirebaseDto._geolocationKey: GeoPoint(
+      MessageFirebaseDto.authorNameKey: nickname,
+      MessageFirebaseDto.messageKey: message ?? '',
+      MessageFirebaseDto.createdKey: FieldValue.serverTimestamp(),
+      MessageFirebaseDto.geolocationKey: GeoPoint(
         location.latitude,
         location.longitude,
       ),
@@ -94,7 +98,7 @@ class ChatRepositoryFirebase implements ChatRepository {
   ChatMessageDto _parseFirebaseDataToLocal(
     QueryDocumentSnapshot<Map<String, dynamic>> snapshot,
   ) {
-    final parsedData = _MessageFirebaseDto.fromMap(snapshot.data());
+    final parsedData = MessageFirebaseDto.fromMap(snapshot.data());
 
     final ChatUserDto author;
     if (parsedData.authorName == _savedLocalName) {
@@ -120,74 +124,6 @@ class ChatRepositoryFirebase implements ChatRepository {
       author: author,
       message: parsedData.message,
       createdDateTime: parsedData.created,
-    );
-  }
-}
-
-class _MessageFirebaseDto {
-  static const String _authorNameKey = 'authorName';
-  static const String _messageKey = 'message';
-  static const String _createdKey = 'created';
-  static const String _geolocationKey = 'geolocation';
-
-  final String authorName;
-  final String message;
-  final DateTime created;
-  final GeoPoint? geolocation;
-
-  _MessageFirebaseDto({
-    required this.authorName,
-    required this.message,
-    required this.created,
-    required this.geolocation,
-  });
-
-  Map<String, dynamic> toMap() {
-    return {
-      _authorNameKey: authorName,
-      _messageKey: message,
-      _createdKey: created.millisecondsSinceEpoch,
-    };
-  }
-
-  factory _MessageFirebaseDto.fromMap(Map<String, dynamic> map) {
-    return _MessageFirebaseDto(
-      authorName: map[_authorNameKey] ?? '',
-      message: map[_messageKey] ?? '',
-      created: (map[_createdKey] as Timestamp).toDate(),
-      geolocation: map[_geolocationKey],
-    );
-  }
-
-  @override
-  String toString() {
-    return '_MessageFirebaseDto(authorName: $authorName, message: $message, created: $created, geolocation: $geolocation)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-
-    return other is _MessageFirebaseDto &&
-        other.authorName == authorName &&
-        other.message == message &&
-        other.created == created;
-  }
-
-  @override
-  int get hashCode => authorName.hashCode ^ message.hashCode ^ created.hashCode;
-
-  _MessageFirebaseDto copyWith({
-    String? authorName,
-    String? message,
-    DateTime? created,
-    GeoPoint? geolocation,
-  }) {
-    return _MessageFirebaseDto(
-      authorName: authorName ?? this.authorName,
-      message: message ?? this.message,
-      created: created ?? this.created,
-      geolocation: geolocation ?? this.geolocation,
     );
   }
 }
