+++ pubspec.yaml
@@ -16,11 +16,18 @@ dependencies:
   firebase_core: ^1.12.0
   cloud_firestore: ^3.1.7
 
+  provider: ^6.0.1
+  flex_color_scheme: ^3.0.1
+
+  #modules
+  core:
+    path: core
+  #
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   flutter_lints: ^1.0.0
-
 flutter:
   uses-material-design: true
